Move UrlHelpers to package utils as it’s going to be used by more than one class
