+++ base/src/main/java/uk/ac/ebi/atlas/model/card/CardModelFactory.java
@@ -6,6 +6,7 @@ import org.springframework.stereotype.Component;
 import uk.ac.ebi.atlas.home.SpeciesSummary;
 import uk.ac.ebi.atlas.model.experiment.Experiment;
 import uk.ac.ebi.atlas.species.Species;
+import uk.ac.ebi.atlas.utils.UrlHelpers;
 
 import java.util.ArrayList;
 import java.util.Collection;
