+++ base/src/test/java/uk/ac/ebi/atlas/configuration/TestConfig.java
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.ComponentScan.Filter;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.web.client.RestTemplate;
-import uk.ac.ebi.atlas.model.card.UrlHelpers;
+import uk.ac.ebi.atlas.utils.UrlHelpers;
 
 @Configuration
 // Enabling component scanning will also load BasePathsConfig, JdbcConfig and SolrConfig, so just using this class as
