+++ gxa/src/main/java/uk/ac/ebi/atlas/model/card/UrlHelpersImpl.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.atlas.model.card;
 
 import org.springframework.stereotype.Component;
+import uk.ac.ebi.atlas.utils.UrlHelpers;
 
 @Component
 class UrlHelpersImpl implements UrlHelpers {}
