+++ sc/src/main/java/uk/ac/ebi/atlas/model/card/UrlHelpersImpl.java
@@ -1,8 +1,7 @@
 package uk.ac.ebi.atlas.model.card;
 
 import org.springframework.stereotype.Component;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-import uk.ac.ebi.atlas.model.experiment.Experiment;
+import uk.ac.ebi.atlas.utils.UrlHelpers;
 
 @Component
 class UrlHelpersImpl implements UrlHelpers {}
